requirements.txt
pytest==8.3.4
requests==2.28.1
mysql-connector-python==8.0.31
allure-pytest==2.8.36
pymysql==1.1.1
Flask==3.0.2
python-dotenv==1.0.1